// types/animeTypes.ts
export interface anime {
  name: string;
  slug: string;
  _id: string;
  status: string;
  season: number;
  favouriteCharacter: string;
  description: string;
  ratingsAverage: number;
  ratingsQuantity: number;
  earnings: number;
  type: string[];
  animeProfile: string;
  animeCover: string;
  createdAt: string;
  id: string;
}